add documentation and annotation to writeback tutorial Signed-off-by: dbulahov <bulahovdenis@gmail.com>
