+++ instance/emf/tutorial/writeback/inlinetable/src/main/resources/data/FACTWB.csv
