+++ instance/emf/tutorial/writeback/inlinetable/src/main/resources/data/L1.csv
