+++ instance/emf/tutorial/writeback/inlinetable/src/main/resources/data/L2.csv
